+++ lib/routes/category_page.dart
@@ -14,7 +14,7 @@ import "package:flow/widgets/category/transactions_info.dart";
 import "package:flow/widgets/flow_card.dart";
 import "package:flow/widgets/general/spinner.dart";
 import "package:flow/widgets/grouped_transaction_list.dart";
-import "package:flow/widgets/home/transactions_date_header.dart";
+import "package:flow/widgets/transactions_date_header.dart";
 import "package:flow/widgets/no_result.dart";
 import "package:flow/widgets/rates_missing_warning.dart";
 import "package:flow/widgets/time_range_selector.dart";